Refactor arrayFunctions to use uint64_t for printArray, update factorialRecursion to use uint64_t, remove unused primeNumberChecker, insertionSort, quickSort, and selectionSort files, add GCD and primeFactors implementations, and create PROGRAMS.md documentation for repository structure.

+++ modules/arrayFunctions.h
@@ -4,9 +4,9 @@
 #include <vector>
 #include <iostream>
 
-void printArray(std::vector<int> &arr)
+void printArray(std::vector<u_int64_t> &arr)
 {
-  for (int num : arr)
+  for (uint64_t num : arr)
   {
     std::cout << num << " ";
   }
